Migrate to SvelteKit 4 | Update Build (#44)
* build(*): Migrate to SvelteKit-4 | Update Build

* build(ci.yml): Add develop branch to test in PRs

* build(ci): Update build server command

BREAKING CHANGE: Build & deploy process

* refactor(footer): Remove link to Galileo
Since Galileo is supposed to be a route, build it separately

* build(ci): change cd build command

* chore: Remove trailing logo.svg

* Update ci.yml

Add PAT mention for gh pages

* Update ci.yml

Add write permisssion to ci script

* Update ci.yml

* Update ci.yml

Add deploy key instead of PAT
