+++ .github/workflows/ci.yml
@@ -1,32 +1,36 @@
 name: Deploy to GitHub pages
 
 on:
- push:
+  push:
     branches:
       - main
- pull_request:
+  pull_request:
     branches:
       - main
+      - develop
+permissions:
+  contents: write
+  
 jobs:
   build:
-
     runs-on: ubuntu-latest
 
     strategy:
       fail-fast: false
       matrix:
-        node-version: [16.x]
+        node-version: [20.x]
     steps:
       - name: Checkout Repo
         uses: actions/checkout@master
 
-      - name: Install & Build
+      - name: Install & Build 🔧
         run: |
           npm install --location=global pnpm
-          npm run build:server
+          pnpm run cd
 
-      - name: Deploy to GitHub Pages
-        uses: JamesIves/github-pages-deploy-action@v4.2.3
+      - name: Deploy to GitHub Pages 🚀
+        uses: JamesIves/github-pages-deploy-action@v4.6.3
         with:
           branch: gh-pages
-          folder: build
+          folder: .svelte-kit/output
+          ssh-key: ${{ secrets.GH_PAGES_DEPLOY_KEY }}
