+++ .gitignore
@@ -1,7 +1,22 @@
-**/node_modules/
+node_modules
 **/*.*.map
 
-.DS_Store
+# Output
+.output
+.vercel
+/.svelte-kit
+/build
 
-**/.svelte-kit
-**/build
+# OS
+.DS_Store
+Thumbs.db
+
+# Env
+.env
+.env.*
+!.env.example
+!.env.test
+
+# Vite
+vite.config.js.timestamp-*
+vite.config.ts.timestamp-*
