+++ DOCS.md
@@ -1,0 +1,15 @@
+# Contribution Guide
+
+## Data Updates
+
+### Homepage
+
+- Legacy: Edit `static/assets/legacy.svg'
+
+### Team
+
+- Management & Lead: `src/lib/data/team.json
+
+### Projects
+
+- Projects: `src/lib/data/projects.json`
