+++ package.json
@@ -1,20 +1,38 @@
 {
-  "name": "korolev",
-  "version": "3.3.2",
-  "author": "Plutoniumm",
-  "owner": "SEDSCelestiaBPGC",
-  "type": "module",
-  "scripts": {
-    "dev": "svelte-kit dev",
-    "build": "node prep && rm -rf build && svelte-kit build --verbose",
-    "build:server": "pnpm i --no-frozen-lockfile && node prep && rm -rf build && svelte-kit build --verbose",
-    "node-sass": "node-sass static/css/global.scss static/css/global.css --output-style compressed"
-  },
-  "devDependencies": {
-    "@sveltejs/adapter-static": "next",
-    "@sveltejs/kit": "next",
-    "node-sass": "^7.0.1",
-    "svelte": "^3.44.0",
-    "svelte-preprocess": "4.7.3"
-  }
+	"name": "korolev",
+	"version": "3.3.2",
+	"author": "Plutoniumm",
+	"owner": "SEDSCelestiaBPGC",
+	"type": "module",
+	"scripts": {
+		"dev": "vite dev",
+		"build": "vite build",
+		"cd": "pnpm i --no-frozen-lockfile && node prep && rm -rf build && vite build",
+		"preview": "vite preview",
+		"lint": "prettier --check . && eslint .",
+		"format": "prettier --write .",
+		"sass": "sass --no-source-map --style=compressed --update static/css/global.scss static/css/global.css",
+		"clean:build": "rm -rf build .svelte-kit",
+		"clean:cache": "rm -rf node_modules .svelte-kit"
+	},
+	"devDependencies": {
+		"@sveltejs/adapter-auto": "^3.0.0",
+		"@sveltejs/adapter-static": "^3.0.4",
+		"@sveltejs/kit": "^2.0.0",
+		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@types/eslint": "^9.6.0",
+		"eslint": "^9.0.0",
+		"eslint-config-prettier": "^9.1.0",
+		"eslint-plugin-svelte": "^2.36.0",
+		"globals": "^15.0.0",
+		"prettier": "^3.1.1",
+		"prettier-plugin-svelte": "^3.1.2",
+		"sass": "^1.77.8",
+		"svelte": "^4.2.7",
+		"svelte-preprocess": "^6.0.2",
+		"vite": "^5.0.3"
+	},
+	"dependencies": {
+		"korolev": "file:"
+	}
 }
