+++ src/app.html
@@ -1,26 +1,42 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en">
+	<head>
+		<meta charset="utf-8" />
+		<link rel="icon" href="/favicon.png" />
+		<meta name="viewport" content="width=device-width, initial-scale=1" />
 
-<head>
-	<meta charset="utf-8" />
-	<link rel="icon" href="/favicon.ico" />
-	<meta name="viewport" content="width=device-width, initial-scale=1" />
+		<link rel="stylesheet" href="/css/atomic.css">
+		<link rel="stylesheet" href="/css/global.css">
 
-	<link rel="stylesheet" href="/css/atomic.css">
-	<link rel="stylesheet" href="/css/global.css">
+		<script src="/js/molecular.js"></script>
 
-	<script src="/js/molecular.js"></script>
+		%sveltekit.head%
+	</head>
+	<body data-sveltekit-preload-data="hover" class="m-0 p-0" xmlns:svg="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
+		<canvas id="star-space" style="z-index: -3;"></canvas>
+		<canvas id="star-field" style="z-index: -1;"></canvas>
+		<div style="display: contents">%sveltekit.body%</div>
+	</body>
 
-	%sveltekit.head%
-</head>
+	<script>
+		const starFieldWidth = 2 * window.innerWidth,
+					starFieldHeight = 2 * window.innerHeight,
+					starField = document.getElementById( "star-field" ),
+					starSpace = document.getElementById( "star-space" ),
+					density = 2 * Math.pow( starFieldHeight * starFieldWidth, 0.25 );
+		const createRandomNodes = ( t, e, s, i, a ) => {
+			a.width = s, a.height = i;
+			let o = a.getContext( "2d" );
+			o.clearRect( 0, 0, s, i );
+			for ( let a = 0;a <= t;a++ ) {
+				o.beginPath();
+				let t = Math.random( a ) * s, d = Math.random( a ) * i;
+				o.arc( t, d, e, 0, 2 * Math.PI ), o.fillStyle = "#fff", o.fill(), o.closePath() }
+				return 0;
+			}
+		createRandomNodes( density, 1.5, starFieldWidth, starFieldHeight, starSpace ), window.innerWidth > 600 && createRandomNodes( density, 2, starFieldWidth, starFieldHeight, starField );
+		</script>
+	<div class="cursor"></div>
+	<script src="/js/cursor.js"></script>
 
-<body class="m-0 p-0" xmlns:svg="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
-	<canvas id="star-space" style="z-index: -3;"></canvas>
-	<canvas id="star-field" style="z-index: -1;"></canvas>
-	<main class="w-100 z-1">%sveltekit.body%</main>
-</body>
-
-<script>const starFieldWidth = 2 * window.innerWidth, starFieldHeight = 2 * window.innerHeight, starField = document.getElementById( "star-field" ), starSpace = document.getElementById( "star-space" ), density = 2 * Math.pow( starFieldHeight * starFieldWidth, 0.25 ); function createRandomNodes( t, e, s, i, a ) { a.width = s, a.height = i; let o = a.getContext( "2d" ); o.clearRect( 0, 0, s, i ); for ( let a = 0;a <= t;a++ ) { o.beginPath(); let t = Math.random( a ) * s, d = Math.random( a ) * i; o.arc( t, d, e, 0, 2 * Math.PI ), o.fillStyle = "#fff", o.fill(), o.closePath() } return 0 } createRandomNodes( density, 1.5, starFieldWidth, starFieldHeight, starSpace ), window.innerWidth > 600 && createRandomNodes( density, 2, starFieldWidth, starFieldHeight, starField );</script>
-<div class="cursor"></div>
-<script src="/js/cursor.js"></script>
 </html>
