+++ src/lib/data/meta.json
@@ -1,1 +1,1 @@
-{"name":"Korolev","author":"Plutoniumm","owner":"SEDSCelestiaBPGC","build":{"current":5,"previous":4,"version":"3.3.2"},"time":1655492344162}
+{"name":"Korolev","author":"Plutoniumm","owner":"SEDSCelestiaBPGC","build":{"current":13,"previous":12,"version":"3.3.2"},"time":1724959194139}
