--- src/routes/__layout.svelte
@@ -1,12 +1,0 @@
-<script>
-	import Navigator from "./components/nav.svelte";
-	import Footer from "./components/footer.svelte";
-</script>
-
-<Navigator />
-
-<main>
-	<slot />
-</main>
-
-<Footer />
